refactor(db): #553 P2 Stage 1 — extract _MIGRATIONS / _SCHEMA to db_migrations.py
Codex consult round 1+2 (data/llm_consults/2026-05-01_p2-db-split-{design,round2}.md)
권고 그대로 — Option C (data-only separation) Stage 1 수행. zero-behavior split.

분리:
- nuri/core/db_migrations.py 신규 (1240 LOC)
  - _SCHEMA, _SCHEMA_VERSION_TABLE, _MIGRATIONS (40 entries) 만 export
  - sqlite3 import 0 (data only)
- nuri/core/db.py: 2872 → 1645 LOC (-1227, 데이터 추출)
  - public API 0 변경, helper 함수 100% 동일
  - from .db_migrations import _MIGRATIONS, _SCHEMA, _SCHEMA_VERSION_TABLE

Codex Stage 1 → Stage 2 안전 신호 6/8 충족:
1. ✓ make test-fast 4476/4476 green
2. ✓ make verify-quick green
3. ✓ tests/core/test_agent_infra.py::test_schema_version_at_40 pass
4. ✓ tests/agents/test_{regime_posterior,hypothesis_registry,
       execution_firewall,sre_incident_agent}.py 전부 pass
5. ✓ tests/conftest.py monkeypatch 영향 0 (DB_PATH binding 변동 X)
6. ✓ ruff clean
7. (manual verify pending — fresh init_db sqlite_master diff)
8. ✓ rg "import sqlite3" nuri 결과 = nuri/core/db.py 1개
   (certification.py 의 import 는 Stage 1 이전부터 존재, scope 외)

Stage 2 (전체 package 분할) 는 별도 PR 예정 — 본 PR ship 후 Codex 권고 safety
gate 1-6 운영 환경 검증 + helper submodule 의 DB_PATH late-binding 강제 후 진행.
